move components to own files
