+++ util.js
@@ -1,13 +1,6 @@
 import React from "react";
-import PropTypes from "prop-types";
-import {Picker, Platform, Text, TouchableOpacity, View, StatusBar} from "react-native";
-import RNPickerSelect from "react-native-picker-select";
-import DateTimePicker from "react-native-modal-datetime-picker";
-import Styles from "./styles";
 import TinyColor from "tinycolor2";
 import RandomColor from "randomcolor";
-import _ from "lodash";
-import {Svg as SVG, Circle} from "react-native-svg";
 
 //handles converting a JS Date object into an ISO date string
 export function toDateString(_date) {
@@ -151,173 +144,10 @@ export function toISO(usDate) {
 let dayInMS = 24 * 60 * 60 * 1000;
 export { dayInMS };
 
-export class Dropdown extends React.Component {
-    static propTypes = {
-        options: PropTypes.arrayOf(PropTypes.shape({
-            label: PropTypes.string,
-            value: PropTypes.string
-        })).isRequired,
-        selectedValue: PropTypes.string,
-        onValueChange: PropTypes.func.isRequired,
-        style: PropTypes.oneOfType([
-            PropTypes.object,
-            PropTypes.arrayOf(PropTypes.object)
-        ])
-    };
-
-    static defaultProps = {
-        style: null
-    };
-
-    constructor(props) {
-        super(props);
-        this.state = {
-            value: props.selectedValue
-        };
-    }
-
-    render() {
-        if (Platform.OS === "android") {
-            return (
-                <Picker style={this.props.style}
-                    selectedValue = {this.state.value}
-                    onValueChange = {value => {
-                        this.setState({value: value});
-                        this.props.onValueChange(value);
-                    }}
-                >
-                    {this.props.options.map(option =>
-                        <Picker.Item key={option.value} label={option.label} value={option.value} />
-                    )}
-                </Picker>
-            );
-        } else {
-            return (
-                <RNPickerSelect style={this.props.style}
-                    items = {this.props.options}
-                    value = {this.state.value}
-                    onValueChange={value => {
-                        this.setState({value: value});
-                        this.props.onValueChange(value);
-                    }}
-                />
-            );
-        }
-    }
-}
-
-//helper class to input time since the timepicker is bizarrely complex
-export class TimeInput extends React.Component {
-    static propTypes = {
-        onValueChange: PropTypes.func.isRequired,
-        value: PropTypes.string
-    };
-
-    constructor(props) {
-        super(props);
-
-        let msIn15Mins = 1000 * 60 * 15;
-        let now = new Date();
-        now.setMilliseconds(Math.ceil(now.getTime() / msIn15Mins) * msIn15Mins);
-
-        this.state = {
-            value: this.props.value ? toDateTime({time: toMilitaryTime(this.props.value)}) : now,
-            open: false
-        };
-    }
-
-    render() {
-        return(
-            <View style={Styles.datetimeContainer}>
-                <TouchableOpacity style={Styles.inputBox}
-                                  onPress = {() => this.setState({open: true})}
-                >
-                    <Text>{toAMPM(toTimeString(this.state.value))}</Text>
-                </TouchableOpacity>
-                <DateTimePicker
-                    date = {this.state.value}
-                    mode = "time"
-                    isVisible={this.state.open}
-                    onConfirm = {time => {
-                        this.setState({
-                            value: time,
-                            open: false
-                        });
-
-                        this.props.onValueChange(toTimeString(time));
-                    }}
-                    onCancel = {() => this.setState({open: false})}
-                />
-            </View>
-        );
-    }
-}
-
 export function randomColor(seed) {
     let color = TinyColor(RandomColor({seed: seed}));
     return {
         hex: color.toHexString(),
         isDark: color.isDark()
     };
-}
-
-export class AppContainer extends React.Component {
-    static propTypes = {
-        style: PropTypes.oneOfType([
-            PropTypes.object,
-            PropTypes.arrayOf(PropTypes.object)
-        ])
-    };
-
-    static defaultProps = {
-        style: null
-    };
-
-    render() {
-        return(
-            <View style={_.flatten([Styles.appContainer, this.props.style])}>
-                <StatusBar barStyle="light-content" backgroundColor="#fff"/>
-                {this.props.children}
-            </View>
-        );
-    }
-}
-
-export class MoreButton extends React.Component {
-    static propTypes = {
-        onPress: PropTypes.func.isRequired
-    };
-
-    static generateCircles() {
-        let circles = [];
-
-        for (let i = 0; i < 4; i++) {
-            let position = (i + 1) * 100 - 50;
-            circles.push(
-                <Circle
-                    key={i.toString()}
-                    cx = {position}
-                    cy = "50"
-                    r = "25"
-                    stroke = "#111"
-                    strokeWidth = "1"
-                    fill = "#111"
-                />
-            );
-        }
-
-        return circles
-    }
-
-    render() {
-        return (
-            <TouchableOpacity style={Styles.moreButton}
-                onPress={this.props.onPress}
-            >
-                <SVG height="100%" width = "100%" viewBox = "0 0 100 300">
-                    {MoreButton.generateCircles()}
-                </SVG>
-            </TouchableOpacity>
-        );
-    }
 }
